make the autherized user who can del and put

+++ app/routers/posts.py
@@ -47,6 +47,10 @@ def delete_post(id:int, db: Session = Depends(get_db) ,current_user:int = Depend
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                         detail=f"Sorry, post with id {id} not found")
     
+    if post_dict.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requests action")
+    
     post_query.delete(synchronize_session=False)
     db.commit()
     return Response(status_code=status.HTTP_204_NO_CONTENT)
@@ -54,11 +58,15 @@ def delete_post(id:int, db: Session = Depends(get_db) ,current_user:int = Depend
 @router.put("/{id}" ,status_code= status.HTTP_202_ACCEPTED, response_model=schemas.PostResponse)
 def update_post(id:int, updated_post:schemas.PostCreate,db: Session = Depends(get_db), current_user:int = Depends(oauth2.get_current_user)):
     post_query = db.query(models.post).filter(models.post.id==id)
-    post = post_query.first()
+    post_dict = post_query.first()
 
-    if post == None:
+    if post_dict == None:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, 
                             detail=f"Sorry, post with id {id} not found")
+    
+    if post_dict.owner_id != current_user.id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN,
+                            detail="Not authorized to perform requests action")
 
     post_query.update(updated_post.dict(),synchronize_session=False)
     db.commit()
